Example binary to render a project dir
